DROID-3424 search bar
